New Project using datatables
Started the project from start with no frills or fancies in it.  DB is the most important thing in it, so I started there.  There are some issues with this one, particularly as you move through the pages, but I think these are only small issues.  There's no designing in the app yet either.  However, there is a searchable and sortable database.